+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 the [PEP 440 version scheme](https://peps.python.org/pep-0440/#version-scheme).
 
+## [Unreleased]
+### Changed
+- License from MIT to LGPLv3.0
+
 ## [v0.1.0-alpha2]
 ### Changed
 - Use tuple instead of ListView for pytrified sequences.